TimeClocker 1.2.3
Finally got rid of all regex from the input processing, and changed it
to a string function that removes the first and last two characters from
each item in the dict. This strips the [''] from the values while adding
it to the list WITHOUT tampering with user input. I am dumb.